Convert advisor home page into multiple components
